chore(@W-9131346): tdalbo/core delivery module (#918)
* chore: update package.json to include sfdc section

This will allow each module to declare the file path and name it should
be shipped as

* Cleanup build file

* WIP - having filesystem perm issues

* WIP: Separating release from building

* WIP:  Copy working, now to implement p4 release

* WIP Add print commit support

* wip:  Core packager working (mostly)

It's requiring the package.json to include a workspaces dir, which
means it'll run on all our individual modules.  This is a problem
because we can't use the `lwc` config in package.json for any of our
modules without it being picked up by core-packager.  This would force
us to include non-core LWR bundled adapters in our core bundle... even if we
aren't exposing it, it's still _not ideal_

* chore(): Updated core packager lwc allowlist

* chore(W-9131346): Maintain both core release scripts

When Core Packager is ready, the script in scripts/release/core.js
can be removed.  For now.. support both.

* chore(): Rename lds-build to core-build

* chore(): PR review cleanup