Add Scale Center Adapter Family and queryMetrics Adapter (#1149)
* Initial commit for Scale Center LDS adapter

Currently running into build issues when running build:raml. Pushing to git to get troubleshooting assistance

* Small changes

* Build bug fix

The build was failing because of the comment at the top of the file. It was unable to parse the RAML. The comment is removed and yarn build:raml works now

* Update return property name

* Add query parameter

Add inputJson parameter to api definition

* Update LDS adapter to use new endpoint and fields

* Update dependencies and RAML

* Add basic success and error unit tests

* Move unit test file

* Update mock response data

* Remove nonfunctional caching ttl

We are not caching yet and this cache TTL was just an artifact of a debugging session when we were setting up the adapter. We will add it back when we add client side caching

* Bump package version

* Add ttl back to luvio.raml

The ttl is necessary for the build to complete even though we have not implemented caching yet

* Add our base URI to connect-base.ts and use it in connect.ts

* Rename endpoint to /scalecenter/metrics/query

* Update unit tests

* Move request path

* Update adapter description

* v0.1.0

* Remove jest target from package.xml

* Remove unnecessary version

* Run Prettier

* Run prettier on karma.conf.js

* Remove .sfdx files

* Move .sfdx from project artifacts to tools artifacts in .gitignore

* upgrade lds-karma version

* upgrade lds-karma and lds-bindings

* update version

* trigger CI

* trigger CI

Co-authored-by: Paymon Teyer <pteyer@users.noreply.github.com>